Do not send invalid dates to datepicker
It turns out that sending invalid year (e.g. 0003)
to the date picker freezes the UI.

This is hotfix only. The true solution would be to
rewrite the schedule form not to use vuex
so it will be actually predictable.

And the true mystery is:
moment("05 0 * * 1,2,3,4,5", "m H D M * Y").format() == "0003-02-01T00:05:00+01:00"

WHY?

fix #324